[US240423] Moving fingerprint methods from MASDevice to MASUser
