+++ MAS/src/main/java/com/ca/mas/foundation/MASDevice.java
@@ -7,24 +7,13 @@
  */
 package com.ca.mas.foundation;
 
-import android.annotation.TargetApi;
-import android.os.Build;
-import android.os.Parcel;
-
 import com.ca.mas.core.MobileSso;
 import com.ca.mas.core.MobileSsoFactory;
 import com.ca.mas.core.conf.ConfigurationManager;
 import com.ca.mas.core.conf.ConfigurationProvider;
 import com.ca.mas.core.context.DeviceIdentifier;
-import com.ca.mas.core.security.DefaultEncryptionProvider;
-import com.ca.mas.core.security.FingerprintListener;
-import com.ca.mas.core.security.LockableKeyStorageProvider;
-import com.ca.mas.core.security.UserNotAuthenticatedException;
 import com.ca.mas.core.store.StorageProvider;
 import com.ca.mas.core.store.TokenManager;
-import com.ca.mas.core.store.TokenStoreException;
-import com.ca.mas.core.token.IdToken;
-import com.ca.mas.core.token.JWTValidation;
 import com.ca.mas.foundation.auth.MASProximityLoginBLEPeripheralListener;
 import com.ca.mas.foundation.notify.Callback;
 import com.ca.mas.foundation.util.FoundationUtil;
@@ -33,19 +22,12 @@ import com.ca.mas.foundation.util.FoundationUtil;
  * <p>The <b>MASDevice</b> class is a local representation of device data.</p>
  */
 public abstract class MASDevice implements Device {
-    private static String ENCRYPTED_KEY = MASDevice.class.getSimpleName() + "EncryptedKey";
-    private static String SUFFIX = MASDevice.class.getSimpleName() + "RandomSuffix";
-    private static LockableKeyStorageProvider mKeyStoreProvider;
     private static MASDevice current;
 
     private MASDevice() {
     }
 
     public static MASDevice getCurrentDevice() {
-        if (Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.M) {
-            mKeyStoreProvider = new LockableKeyStorageProvider(SUFFIX);
-        }
-
         if (current == null) {
             current = new MASDevice() {
                 @Override
@@ -97,138 +79,19 @@ public abstract class MASDevice implements Device {
                 public void stopAsBluetoothPeripheral() {
                     MobileSsoFactory.getInstance().stopBleSessionSharing();
                 }
-
-                @Override
-                @TargetApi(23)
-                public void lockSession(MASCallback<Void> callback) {
-                    MASUser currentUser = MASUser.getCurrentUser();
-                    if (currentUser == null || !currentUser.isAuthenticated()) {
-                        callback.onError(new MASException("No currently authenticated user."));
-                    } else if (isSessionLocked()) {
-                        callback.onError(new MASException("Device is already locked."));
-                    } else {
-                        // Remove access and refresh tokens
-//                        MssoContext context = MssoContext.newContext();
-//                        context.clearAccessToken();
-                        TokenManager keyChainManager = createTokenManager();
-
-                        // Remove persisted user profile
-//                        try {
-//                            keyChainManager.deleteUserProfile();
-//                        } catch (TokenStoreException e) {
-//                            callback.onError(e);
-//                        }
-
-                        // Validate that the ID token is not expired
-                        IdToken idToken = keyChainManager.getIdToken();
-                        boolean isTokenExpired = JWTValidation.isIdTokenExpired(idToken);
-                        if (!isTokenExpired) {
-                            // Move the ID token from the Keychain to the fingerprint protected shared Keystore
-                            Parcel idTokenParcel = Parcel.obtain();
-                            idToken.writeToParcel(idTokenParcel, 0);
-                            byte[] idTokenBytes = idTokenParcel.marshall();
-
-                            DefaultEncryptionProvider encryptionProvider = new DefaultEncryptionProvider(MAS.getContext(), mKeyStoreProvider);
-                            byte[] encryptedData = encryptionProvider.encrypt(idTokenBytes);
-                            // Save the encrypted token
-                            try {
-                                keyChainManager.saveSecureIdToken(encryptedData);
-                            } catch (TokenStoreException e) {
-                                callback.onError(new MASException("Could not save encrypted ID token."));
-                            }
-
-                            // Remove the unencrypted token
-                            try {
-                                keyChainManager.deleteIdToken();
-                            } catch (TokenStoreException e) {
-                                callback.onError(new MASException("Failed to delete ID token."));
-                            }
-
-                            mKeyStoreProvider.lock();
-                            idTokenParcel.recycle();
-
-                            callback.onSuccess(null);
-                        } else {
-                            callback.onError(new MASException("ID token is expired."));
-                        }
-                    }
-                }
-
-                @Override
-                @TargetApi(23)
-                public void unlockSession(FingerprintListener listener, MASCallback<Void> callback) {
-//                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-//                    FingerprintManager manager = context.getSystemService(FingerprintManager.class);
-//                    if (ContextCompat.checkSelfPermission(context, Manifest.permission.USE_FINGERPRINT) != PackageManager.PERMISSION_GRANTED) {
-//                        // Fingerprint permission not provided
-//                    } else if (!manager.isHardwareDetected()) {
-//                        // Fingerprint hardware not provided
-//                    } else if (!manager.hasEnrolledFingerprints()) {
-//                        // User hasn't enrolled any fingerprints to authenticate with
-//                    } else {
-////                            manager.authenticate();
-//                    }
-//                }
-                    if (isSessionLocked()) {
-                        TokenManager keyChainManager = createTokenManager();
-
-                        // Unlock the ID token from the Keystore and places the decrypted ID token back to the Keychain
-                        byte[] secureIdToken = keyChainManager.getSecureIdToken();
-
-                        DefaultEncryptionProvider encryptionProvider = new DefaultEncryptionProvider(MAS.getContext(), mKeyStoreProvider);
-                        // Read the decrypted data, reconstruct it as a Parcel, then as an IdToken
-                        Parcel parcel = Parcel.obtain();
-                        try {
-                            // Decrypt the encrypted ID token
-                            byte[] decryptedData = encryptionProvider.decrypt(secureIdToken);
-                            parcel.unmarshall(decryptedData, 0, decryptedData.length);
-                            parcel.setDataPosition(0);
-
-                            IdToken token = IdToken.CREATOR.createFromParcel(parcel);
-                            try {
-                                // Save the unlocked ID token
-                                keyChainManager.saveIdToken(token);
-                            } catch (TokenStoreException e) {
-                                callback.onError(new MASException("Failed to save ID token."));
-                            }
-
-                            try {
-                                // Remove the locked ID token
-                                keyChainManager.deleteSecureIdToken();
-                            } catch (TokenStoreException e) {
-                                callback.onError(new MASException("Failed to delete encrypt ID token."));
-                            }
-
-                            // Delete the previously generated key after successfully decrypting
-                            mKeyStoreProvider.removeKey(DefaultEncryptionProvider.KEY_ALIAS);
-                            // Indicate the device is unlocked
-                            callback.onSuccess(null);
-                        } catch (Exception e) {
-                            if (e instanceof UserNotAuthenticatedException
-                                    || e.getCause() instanceof android.security.keystore.UserNotAuthenticatedException) {
-                                // Listener activity to trigger fingerprint
-                                listener.triggerDeviceUnlock();
-                            }
-                        }
-                    } else {
-                        callback.onError(new MASException("Device is already unlocked."));
-                    }
-                }
-
-                @Override
-                @TargetApi(23)
-                public boolean isSessionLocked() {
-                    TokenManager keyChainManager = createTokenManager();
-                    return keyChainManager.getSecureIdToken() != null;
-                }
             };
         }
         return current;
     }
 
-    private static TokenManager createTokenManager() {
+    static StorageProvider createStorageProvider() {
         ConfigurationProvider configurationProvider = ConfigurationManager.getInstance().getConnectedGatewayConfigurationProvider();
         StorageProvider storageProvider = new StorageProvider(MAS.getContext(), configurationProvider);
-        return storageProvider.createTokenManager();
+        return storageProvider;
     }
+
+    static TokenManager createTokenManager() {
+        return createStorageProvider().createTokenManager();
+    }
+
 }
