+++ MAS/src/main/java/com/ca/mas/foundation/MASUser.java
@@ -7,9 +7,11 @@
  */
 package com.ca.mas.foundation;
 
+import android.annotation.TargetApi;
 import android.content.AsyncTaskLoader;
 import android.graphics.Bitmap;
 import android.os.Handler;
+import android.os.Parcel;
 import android.support.annotation.NonNull;
 import android.util.Log;
 
@@ -18,8 +20,16 @@ import com.ca.mas.core.MobileSso;
 import com.ca.mas.core.MobileSsoFactory;
 import com.ca.mas.core.error.MAGError;
 import com.ca.mas.core.http.MAGResponse;
+import com.ca.mas.core.security.DefaultEncryptionProvider;
+import com.ca.mas.core.security.FingerprintListener;
+import com.ca.mas.core.security.LockableKeyStorageProvider;
+import com.ca.mas.core.security.UserNotAuthenticatedException;
+import com.ca.mas.core.store.OAuthTokenContainer;
 import com.ca.mas.core.store.StorageProvider;
 import com.ca.mas.core.store.TokenManager;
+import com.ca.mas.core.store.TokenStoreException;
+import com.ca.mas.core.token.IdToken;
+import com.ca.mas.core.token.JWTValidation;
 import com.ca.mas.core.util.Functions;
 import com.ca.mas.foundation.notify.Callback;
 import com.ca.mas.foundation.util.FoundationUtil;
@@ -51,6 +61,10 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
+import static android.provider.ContactsContract.CommonDataKinds.StructuredName.SUFFIX;
+import static com.ca.mas.foundation.MASDevice.createStorageProvider;
+import static com.ca.mas.foundation.MASDevice.createTokenManager;
+
 /**
  * <p>The <b>MASUser</b> interface is a core interface used to represent a user in the system,
  * as well as an Identity Management definition used in the interaction with SCIM.</p>
@@ -61,12 +75,13 @@ import java.util.NoSuchElementException;
  * "urn:ietf:params:scim:schemas:core:2.0:User". See the <a href="https://tools.ietf.org/html/rfc7643#section-8.2">Full User Representation</a> for the complete format.</p>
  */
 public abstract class MASUser implements MASTransformable, MASMessenger, MASUserIdentity, ScimUser {
-
+    private static final String STRING_DEVICE_ALREADY_UNLOCKED = "The device is already unlocked.";
+    private static final String STRING_FAILED_TO_DELETE_SECURE_TOKEN = "Failed to delete encrypt ID token.";
     private static List<UserRepository> userRepositories = new ArrayList<>();
 
     static {
-        //For Social Login, the scim endpoint should failed and fallback to /userinfo
-        //Try to get the scim profile first then the userinfo, order in the list is important
+        // For Social Login, the SCIM endpoint should failed and fallback to /userinfo.
+        // Try to get the SCIM profile first followed by the userinfo, this order is important.
         userRepositories = new ArrayList<>();
         userRepositories.add(new ScimUserRepository());
         userRepositories.add(new UserInfoRepository());
@@ -137,13 +152,17 @@ public abstract class MASUser implements MASTransformable, MASMessenger, MASUser
         return current;
     }
 
+    @TargetApi(23)
+    private static LockableKeyStorageProvider initializeKeyStoreProvider() {
+        return new LockableKeyStorageProvider(SUFFIX);
+    }
+
     private static MASUser createMASUser() {
 
         return new MASUser() {
-
             private TokenManager tokenManager = new StorageProvider(MAS.getContext()).createTokenManager();
-
             private ScimUser scimUser = getLocalUserProfile();
+            private LockableKeyStorageProvider mKeyStoreProvider = initializeKeyStoreProvider();
 
             @Override
             public boolean isAuthenticated() {
@@ -497,6 +516,137 @@ public abstract class MASUser implements MASTransformable, MASMessenger, MASUser
                 }
                 return user;
             }
+
+            @Override
+            @TargetApi(23)
+            public void lockSession(MASCallback<Void> callback) {
+                MASUser currentUser = MASUser.getCurrentUser();
+                if (currentUser == null || !currentUser.isAuthenticated()) {
+                    callback.onError(new MASException("No currently authenticated user."));
+                } else if (isSessionLocked()) {
+                    callback.onError(new MASException("Device is already locked."));
+                } else {
+                    // Remove access and refresh tokens
+                    StorageProvider storageProvider = createStorageProvider();
+                    OAuthTokenContainer container = storageProvider.createOAuthTokenContainer();
+                    container.clearAll();
+
+                    // Validate that the ID token is not expired
+                    TokenManager keyChainManager = createTokenManager();
+                    IdToken idToken = keyChainManager.getIdToken();
+                    boolean isTokenExpired = JWTValidation.isIdTokenExpired(idToken);
+                    if (!isTokenExpired) {
+                        // Move the ID token from the Keychain to the fingerprint protected shared Keystore
+                        Parcel idTokenParcel = Parcel.obtain();
+                        idToken.writeToParcel(idTokenParcel, 0);
+                        byte[] idTokenBytes = idTokenParcel.marshall();
+
+                        DefaultEncryptionProvider encryptionProvider = new DefaultEncryptionProvider(MAS.getContext(), mKeyStoreProvider);
+                        byte[] encryptedData = encryptionProvider.encrypt(idTokenBytes);
+                        // Save the encrypted token
+                        try {
+                            keyChainManager.saveSecureIdToken(encryptedData);
+                        } catch (TokenStoreException e) {
+                            callback.onError(new MASException("Could not save encrypted ID token."));
+                        }
+
+                        // Remove the unencrypted token
+                        try {
+                            keyChainManager.deleteIdToken();
+                        } catch (TokenStoreException e) {
+                            callback.onError(new MASException("Failed to delete ID token."));
+                        }
+
+                        mKeyStoreProvider.lock();
+                        idTokenParcel.recycle();
+
+                        callback.onSuccess(null);
+                    } else {
+                        callback.onError(new MASException("ID token is expired."));
+                    }
+                }
+            }
+
+            @Override
+            @TargetApi(23)
+            public void unlockSession(FingerprintListener listener, MASCallback<Void> callback) {
+//                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+//                    FingerprintManager manager = context.getSystemService(FingerprintManager.class);
+//                    if (ContextCompat.checkSelfPermission(context, Manifest.permission.USE_FINGERPRINT) != PackageManager.PERMISSION_GRANTED) {
+//                        // Fingerprint permission not provided
+//                    } else if (!manager.isHardwareDetected()) {
+//                        // Fingerprint hardware not provided
+//                    } else if (!manager.hasEnrolledFingerprints()) {
+//                        // User hasn't enrolled any fingerprints to authenticate with
+//                    } else {
+////                            manager.authenticate();
+//                    }
+//                }
+                if (isSessionLocked()) {
+                    TokenManager keyChainManager = createTokenManager();
+
+                    // Unlock the ID token from the Keystore and places the decrypted ID token back to the Keychain
+                    byte[] secureIdToken = keyChainManager.getSecureIdToken();
+
+                    DefaultEncryptionProvider encryptionProvider = new DefaultEncryptionProvider(MAS.getContext(), mKeyStoreProvider);
+                    // Read the decrypted data, reconstruct it as a Parcel, then as an IdToken
+                    Parcel parcel = Parcel.obtain();
+                    try {
+                        // Decrypt the encrypted ID token
+                        byte[] decryptedData = encryptionProvider.decrypt(secureIdToken);
+                        parcel.unmarshall(decryptedData, 0, decryptedData.length);
+                        parcel.setDataPosition(0);
+
+                        IdToken token = IdToken.CREATOR.createFromParcel(parcel);
+                        try {
+                            // Save the unlocked ID token
+                            keyChainManager.saveIdToken(token);
+                        } catch (TokenStoreException e) {
+                            callback.onError(new MASException("Failed to save ID token."));
+                        }
+
+                        try {
+                            // Remove the locked ID token
+                            keyChainManager.deleteSecureIdToken();
+                        } catch (TokenStoreException e) {
+                            callback.onError(new MASException("Failed to delete encrypt ID token."));
+                        }
+
+                        // Delete the previously generated key after successfully decrypting
+                        mKeyStoreProvider.removeKey(DefaultEncryptionProvider.KEY_ALIAS);
+                        // Indicate the device is unlocked
+                        callback.onSuccess(null);
+                    } catch (Exception e) {
+                        if (e instanceof UserNotAuthenticatedException
+                                || e.getCause() instanceof android.security.keystore.UserNotAuthenticatedException) {
+                            // Listener activity to trigger fingerprint
+                            listener.triggerDeviceUnlock();
+                        }
+                    }
+                } else {
+                    callback.onError(new MASException(STRING_DEVICE_ALREADY_UNLOCKED));
+                }
+            }
+
+            @Override
+            public boolean isSessionLocked() {
+                TokenManager keyChainManager = createTokenManager();
+                return keyChainManager.getSecureIdToken() != null;
+            }
+
+            @Override
+            public void removeSessionLock(MASCallback<Void> callback) {
+                if (!isSessionLocked()) {
+                    callback.onError(new MASException(STRING_DEVICE_ALREADY_UNLOCKED));
+                }
+
+                try {
+                    TokenManager keyChainManager = createTokenManager();
+                    keyChainManager.deleteSecureIdToken();
+                } catch (TokenStoreException e) {
+                    callback.onError(new MASException(STRING_FAILED_TO_DELETE_SECURE_TOKEN));
+                }
+            }
         };
     }
 
@@ -508,7 +658,6 @@ public abstract class MASUser implements MASTransformable, MASMessenger, MASUser
      * @param callback The Callback that receives the results. On a successful completion, the user
      *                 will be logout from the Application.
      */
-
     public abstract void logout(final MASCallback<Void> callback);
 
     /**
@@ -528,4 +677,34 @@ public abstract class MASUser implements MASTransformable, MASMessenger, MASUser
      *                 available via {@link MASUser#getCurrentUser()} will be updated with the new information.
      */
     public abstract void requestUserInfo(MASCallback<Void> callback);
+
+    /**
+     * Locks the current device.
+     * This will remove the access and refresh tokens, as well as the user profile.
+     * The ID token will then be moved to the fingerprint protected shared KeyStore.
+     */
+    @TargetApi(23)
+    public abstract void lockSession(MASCallback<Void> callback);
+
+    /**
+     * Triggers the OS level unlock and captures the unlock result.
+     * Unlocks the ID_TOKEN from the KeyStore and places it back into the keychain,
+     * removes the locked ID_TOKEN, and then indicates if the device is unlocked.
+     */
+    @TargetApi(23)
+    public abstract void unlockSession(FingerprintListener listener, MASCallback<Void> callback);
+
+    /**
+     * Checks to see if the device has a locked ID token.
+     *
+     * @return true if there's a secured ID token
+     */
+    @TargetApi(23)
+    public abstract boolean isSessionLocked();
+
+    /**
+     * Removes the secured ID token.
+     */
+    @TargetApi(23)
+    public abstract void removeSessionLock(MASCallback<Void> callback);
 }
