+++ MAS/src/main/java/com/ca/mas/identity/user/UserIdentityManager.java
@@ -15,6 +15,7 @@ import android.text.TextUtils;
 
 import com.ca.mas.core.error.MAGError;
 import com.ca.mas.core.http.MAGResponse;
+import com.ca.mas.core.security.FingerprintListener;
 import com.ca.mas.foundation.MAS;
 import com.ca.mas.foundation.MASCallback;
 import com.ca.mas.foundation.MASException;
@@ -422,7 +423,6 @@ public class UserIdentityManager {
             @Override
             public void sendMessage(MASMessage message, MASUser user, MASCallback<Void> callback) {
                 throw new UserNotAuthenticatedException();
-
             }
 
             @Override
@@ -465,6 +465,25 @@ public class UserIdentityManager {
                 return UserIdentityManager.getInstance().getUserThumbnailImage(this);
             }
 
+            @Override
+            public void lockSession(MASCallback<Void> callback) {
+                throw new UserNotAuthenticatedException();
+            }
+
+            @Override
+            public void unlockSession(FingerprintListener listener, MASCallback<Void> callback) {
+                throw new UserNotAuthenticatedException();
+            }
+
+            @Override
+            public boolean isSessionLocked() {
+                return false;
+            }
+
+            @Override
+            public void removeSessionLock(MASCallback<Void> callback) {
+                throw new UserNotAuthenticatedException();
+            }
         };
     }
